+++ src/consts/index.ts
@@ -86,5 +86,5 @@ export const CHEERUP_EVEN: TIPS_FOR_TODAY_EVEN = {
   28: "友との約束、そして、なにより自分自身との約束を守れる自分になりましょう",
   29: "「栄光」とは「忍耐」です。大事なのは粘り強さです",
   30: "最後までリングに立っていたら、勝負は勝ちです",
-  31: "困っている人人の心に寄り添って、悩みを引き出せる強さを持ちましょう",
+  31: "困っている人の心に寄り添って、悩みを引き出せる強さを持ちましょう",
 };
