+++ src/main.ts
@@ -1,5 +1,13 @@
 import { createApp } from 'vue'
-import './style.css'
 import App from './App.vue'
+import { createPinia } from 'pinia'
+import router from './router'
+
+import './style.css'
+
+const app = createApp(App)
+app.use(createPinia())
+app.use(router)
 
 createApp(App).mount('#app')
+
